move method to choice_providers module
